refactor: MemberRepository를 정상적으로 DI받도록 변경
